linter updates and rename a index.js
