+++ moped-editor/src/views/auth/LoginView.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Link as RouterLink, useNavigate } from "react-router-dom";
+import { Link as RouterLink } from "react-router-dom";
 import * as Yup from "yup";
 import { Formik } from "formik";
 import {
@@ -28,7 +28,6 @@ const useStyles = makeStyles(theme => ({
 
 const LoginView = () => {
   const classes = useStyles();
-  const navigate = useNavigate();
 
   const { login } = useUser();
